+++ demos/jobslow/template/jobinfo.py
@@ -1,0 +1,8 @@
+"""API and resource definitions."""
+
+
+def mock():
+    return "Example"
+
+
+can_resume = True