+++ demos/jobslow/template/submit.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+parman-sbatch-wait run